+++ dbmodel/schemas/main/common/functions/fct/gw_fct_setnodefromarc.sql
@@ -65,6 +65,7 @@ v_node_id text;
 v_selection_mode text;
 v_id text;
 v_querytext text;
+v_error_context text;
 
 BEGIN
 
@@ -226,6 +227,12 @@ EXECUTE 'SELECT gw_fct_getmessage($${"data":{"function":"2118", "fid":"116", "cr
 		     ',"data":{ "info":'||v_result_info||','||
 				'"point":'||v_result_point||'}}'||
 	    '}')::json, 2118, null, null, null);
+
+	-- Exception handling
+	EXCEPTION WHEN OTHERS THEN
+	GET STACKED DIAGNOSTICS v_error_context = PG_EXCEPTION_CONTEXT;
+	RETURN json_build_object('status', 'Failed', 'NOSQLERR', SQLERRM, 'message', json_build_object('level', right(SQLSTATE, 1), 'text', SQLERRM), 'SQLSTATE', SQLSTATE, 'SQLCONTEXT', v_error_context)::json;
+
 
 END;
 $BODY$